Disable Gem signing due to private key requirement on different env.
Specifically, if the private key used for signing the gem is not present during
the `bundle install` phase on Travis, the whole build will break.
I still have to investigate this issue, as I can't figure out yet how to avoid
publishing the private key in the repo and still being able to run the tests.
So, for the moment I will just disable signing, which makes me feel like a
miserable human being.