Re-enable gem signing, with some help from the context.
Now, private signing key path is read from the `RUBYGEMS_SIGNING_KEY_PATH`
environment variable. If it's not set, Ruby will set it to `nil` which means
the gem won't be signed.
This is useful (actually, mandatory) in Travis, since the gem under development
is only installed to test the code and the like, not to be actually released
(at least for now). Though, this improvement might be useful as well the moment
the gem will be released directly from Travis, as the signing key might be
stored on different locations on the dev machines and the Travis build slave.